@@ -8,7 +8,7 @@
     "build": "cross-env  NODE_ENV=production NODE_TS=true NODE_MODE=server webpack --mode production --config webpack-config/webpack.prod.js",
     "build:plugin": "webpack --mode production --plugin=test --display-error-details --config webpack-config/webpack.prod.js",
     "server": "./scripts/start.sh",
-    "start": "cross-env NODE_TS=true NODE_MODE=server webpack serve --env dev --host 0.0.0.0 --config webpack-config/webpack.server.js",
+    "start": "cross-env TS_NODE_PROJECT=\"tsconfig.json\" webpack serve --env dev --host 0.0.0.0 --config webpack-config/webpack.server.ts",
     "lint": "eslint --quiet --fix --ext .ts,.tsx src",
     "prettier": "prettier --write '**/*.{js,jsx,tsx,ts,less,md,json}'",
     "prepare": "husky install",
@@ -50,7 +50,7 @@
     "dva-model-creator": "^0.4.3",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^6.2.10",
+    "fork-ts-checker-webpack-plugin": "^8.0.0",
     "friendly-errors-webpack-plugin": "^1.7.0",
     "gulp": "^4.0.2",
     "history": "^5.3.0",
@@ -84,7 +84,7 @@
     "style-loader": "^1.2.1",
     "three": "^0.146.0",
     "ts-loader": "^8.1.0",
-    "typescript": "^4.2.3",
+    "typescript": "^5.1.3",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "umi-request": "^1.3.9",
     "useTabPanel": "link:@mui/base/useTabPanel",
@@ -102,12 +102,17 @@
     "@swc/helpers": "^0.4.11",
     "@types/classnames": "^2.2.11",
     "@types/css-mediaquery": "^0.1.1",
+    "@types/css-tree": "^2.3.1",
     "@types/event-stream": "^4.0.0",
+    "@types/friendly-errors-webpack-plugin": "^0.1.4",
+    "@types/gulp": "^4.0.11",
     "@types/gulp-filter": "^3.0.34",
+    "@types/highlight.js": "^10.1.0",
     "@types/isomorphic-fetch": "^0.0.35",
     "@types/less": "^3.0.2",
     "@types/lodash": "^4.14.167",
-    "@types/node": "^15.6.1",
+    "@types/markdown-it": "^12.2.3",
+    "@types/node": "^20.3.1",
     "@types/qs": "^6.9.7",
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.6",
@@ -116,6 +121,7 @@
     "@types/three": "^0.146.0",
     "@types/uuid": "^9.0.1",
     "@types/vinyl": "^2.0.7",
+    "@types/webpack": "^5.28.1",
     "@typescript-eslint/eslint-plugin": "^5.10.0",
     "@typescript-eslint/parser": "^5.10.0",
     "ansi-colors": "^4.1.3",
@@ -153,11 +159,10 @@
     "husky": "^8.0.1",
     "lerna": "^4.0.0",
     "less-loader": "^8.0.0",
-    "less-vars-to-js": "^1.3.0",
     "lint-staged": "^13.0.2",
     "loader-utils": "^3.2.0",
     "markdown-it": "^13.0.1",
-    "mini-css-extract-plugin": "^1.4.0",
+    "mini-css-extract-plugin": "^2.7.6",
     "minimist": "^1.2.6",
     "monaco-editor-webpack-plugin": "^7.0.1",
     "opn": "^6.0.0",
@@ -172,13 +177,15 @@
     "stylelint-config-standard": "^29.0.0",
     "stylelint-less": "^1.0.6",
     "swc-loader": "^0.2.3",
+    "ts-node": "^10.9.1",
+    "tsconfig-paths": "^4.2.0",
     "turbo": "^1.3.1",
     "vinyl": "^2.2.1",
     "vinyl-fs": "^3.0.3",
     "webpack": "^5.28.0",
-    "webpack-cli": "^4.6.0",
+    "webpack-cli": "^5.1.4",
     "webpack-dev-middleware": "^5.3.3",
-    "webpack-dev-server": "^3.11.2",
+    "webpack-dev-server": "^4.15.1",
     "webpack-hot-middleware": "^2.25.2",
     "webpack-merge": "^4.2.2"
   },
