@@ -1,18 +1,33 @@
-const path = require('path');
-const rules = require('./lib/rule.js');
-const getPlugins = require('./lib/plugins');
-const proxy = require('./lib/proxy');
-const minimist = require('minimist');
-const ReactRefreshWebpackPlugin = require('@pmmmwh/react-refresh-webpack-plugin');
-const webpack = require('webpack');
+import  path from 'path';
+import rules from './lib/rule';
+import getPlugins from './lib/plugins';
+import proxy, { DEV_SERVER_ENV } from './lib/proxy';
+import  minimist from 'minimist';
+import  ReactRefreshWebpackPlugin from '@pmmmwh/react-refresh-webpack-plugin';
+import  webpack from 'webpack';
+// import csstree from 'css-tree';
+// let config = {};
 
-let config = {};
-
+// console.log(csstree);
 const isDevelopment = process.env.NODE_ENV !== 'production';
 
-if (process.env.NODE_MODE !== 'plugin') {
-  config = require('../.gents.ts');
-}
+// const ast = csstree.parse('.example { world: "!" }');
+
+// traverse AST and modify it
+// csstree.walk(ast, (node) => {
+//   console.log(node ,'node' +
+//     '');
+//   if (node.type === 'ClassSelector' && node.name === 'example') {
+//     node.name = 'hello';
+//   }
+// });
+
+// generate CSS from AST
+// console.log(csstree.generate(ast));
+
+// if (process.env.NODE_MODE !== 'plugin') {
+//   config = require('../.gents.ts');
+// }
 const args = minimist(process.argv.slice(2), {
   string: [
     'env',
@@ -21,8 +36,14 @@ const args = minimist(process.argv.slice(2), {
   ]
 });
 
+const env = args.env as DEV_SERVER_ENV;
+
+function isNilBoolean<T>(argument: T | false): argument is T {
+  return argument !== false;
+}
+
 // 生成webpack文件
-async function renderWebpack(execEnv = 'webpack') {
+async function renderWebpack(execEnv = 'webpack'): Promise<webpack.Configuration> {
   const plugins = await getPlugins();
   
   return {
@@ -30,13 +51,15 @@ async function renderWebpack(execEnv = 'webpack') {
     mode: 'development',
     resolve: {
       extensions: ['.js', '.jsx', '.ts', '.tsx', '.json'],
-      alias: config.alias
+      alias: {
+        '@': path.resolve(__dirname, '..', 'src'),
+      }
     },
     entry: {
       app: [
         path.resolve(__dirname, '..', 'src', 'index.tsx'),
         execEnv === 'bash' && 'webpack-hot-middleware/client?path=/__webpack_hmr&reload=true',
-      ].filter(app => app),
+      ].filter(isNilBoolean),
       'editor.worker': 'monaco-editor/esm/vs/editor/editor.worker.js',
       'json.worker': 'monaco-editor/esm/vs/language/json/json.worker',
       'css.worker': 'monaco-editor/esm/vs/language/css/css.worker',
@@ -72,27 +95,33 @@ async function renderWebpack(execEnv = 'webpack') {
       ...plugins,
       isDevelopment && new ReactRefreshWebpackPlugin(),
       execEnv === 'bash'  && new webpack.HotModuleReplacementPlugin(),
-    ].filter(Boolean),
+    ].filter(isNilBoolean),
     devServer: {
       host: args.host,
       hot: true,
-      contentBase: path.join(__dirname, '..', 'public'), // 用于指定资源目录
+      // contentBase: path.join(__dirname, '..', 'public'), // 用于指定资源目录
       compress: true,
       port: args.port,
       historyApiFallback: true,
-      clientLogLevel: 'silent',
+      // clientLogLevel: 'silent',
       // before: function (app, server, compiler) {
       //   app.get("/some/path", function (req, res) {
       //     res.json({ custom: "response" });
       //   });
       // }
-      proxy: proxy[args.env],
+      proxy: proxy[env],
     },
     externals: {
       './cptable': 'var cptable',
       '../xlsx.js': 'var _XLSX'
-    }
-  };
+    },
+    watchOptions: {
+      // for some systems, watching many files can result in a lot of CPU or memory usage
+      // https://webpack.js.org/configuration/watch/#watchoptionsignored
+      // don't use this pattern, if you have a monorepo with linked packages
+      ignored: /node_modules/,
+    },
+  } as webpack.Configuration;
 }
 
-module.exports = renderWebpack;
+export default renderWebpack;
