@@ -1,25 +1,7 @@
 import React from 'react';
-import { Theme } from '@mui/material/styles';
-import createStyles from '@mui/styles/createStyles';
-import makeStyles from '@mui/styles/makeStyles';
 import Snackbar from '@mui/material/Snackbar';
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    bottom1: {
-      marginBottom: theme.spacing(0),
-    },
-    bottom2: {
-      marginBottom: theme.spacing(8),
-    },
-    bottom3: {
-      marginBottom: theme.spacing(16),
-    },
-  }),
-);
-
 export default function SnackbarUI({ index, toast }: any) {
-  const classes = useStyles('');
   // const handleClose = (event: React.SyntheticEvent | React.MouseEvent, reason?: string) => {
   //     if (reason === 'clickaway') {
   //         return;
@@ -31,7 +13,7 @@ export default function SnackbarUI({ index, toast }: any) {
   return (
     <div>
       <Snackbar
-        className={classes[`bottom${index}`]}
+        style={{ marginTop: index * 8 }}
         anchorOrigin={{
           vertical: 'bottom',
           horizontal: 'left',
