@@ -2,7 +2,9 @@ import fetch from './fetch';
 import { FetchRequestConfig, FetchResponse } from './fetch/types';
 import { ImageListParams } from '@/services/docker/image';
 
-const codeMessage = {
+const codeMessage: {
+  [key: string]: string;
+} = {
   200: '服务器成功返回请求的数据。',
   201: '新建或修改数据成功。',
   202: '一个请求已经进入后台排队（异步任务）。',
