@@ -15,6 +15,7 @@ interface Window {
   //在这里声明xxx之后就能在文件中 window.xxx这样调用了
   luckysheet?: any;
   p5: Constructable<P5>;
+  Babel: any;
 }
 
 declare module '*.svg';
@@ -31,6 +32,8 @@ declare module '*.json';
 
 // declare var process: any;
 // declare var require: any;
-declare var ENV_PATH: string;
+declare let ENV_PATH: string;
 
-declare var MODELS_PATH: Array<string>;
+declare let MODELS_PATH: Array<string>;
+
+declare module 'less-vars-to-js';
